Move LambdaLink to the core atoms directory
