+++ opencog/atoms/core/LambdaLink.cc
