+++ opencog/atoms/core/LambdaLink.h
